Finassistant/pyproject.toml
[tool.poetry]
name = "finassistant"
version = "0.1.0"
description = ""
authors = ["Georgy <georgy111111222222@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
django-jazzmin = "^3.0.1"
drf-yasg = "^1.21.10"


[build-system]
requires = ["poetry-core>=1.8"]
build-backend = "poetry.core.masonry.api"
